🔨 [#096] - Extract duplicated code into LeaveGuards trait 🧹
- 🔨 Moved guardIsPending() into LeaveGuards trait with generic message
- 🔨 Added buildLeaveAttributes() to LeaveGuards eliminating Leave::create() duplication
- 🔨 Added resolveLeaveContext() to LeaveGuards eliminating preamble duplication
- 🔨 Simplified RegisterDirectLeaveAction using shared trait methods
- 🔨 Simplified RegisterLeaveRequestAction using shared trait methods
- 🔨 Removed private guardIsPending() from ApproveLeaveAction and RejectLeaveAction
- 🔨 Added use LeaveGuards to RejectLeaveAction
- 🐛 Changed gte:start_date to after_or_equal:start_date in RegisterDirectLeaveRequest
